src/PlayGame/requirements.txt
# Wheel of Fortune Solve Advisor Dependencies

# Google Gemini API for AI-powered puzzle analysis
google-generativeai>=0.3.0

# Optional: for loading API keys from .env file
python-dotenv>=1.0.0

